requirements.txt
# Dasar Komunikasi
paho-mqtt
pyyaml
numpy
